Switched to use os.Exit function for proper behavior on fatal errors. Also adjusted logging

+++ config/reader_test.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"os"
+	"os/exec"
 	"testing"
 )
 
@@ -46,31 +47,52 @@ func TestReadConfig(T *testing.T) {
 		T.Errorf("Incorrect value in configuration!")
 	}
 
-	//now munge the data and try again. We should get a panic
-	func() {
-		defer func() {
-			if recover() == nil {
-				T.Errorf("Failed to panic on bad data")
-			}
-		}()
-		bitties = bitties[1:]
+}
+
+//Thank Andrew Garrand :)
+func TestReaderExits(T *testing.T) {
+	testData := map[string]interface{}{
+		"testInt":    21,
+		"testString": "Hello",
+		"testBool":   true,
+		"testFloat":  1.23,
+		"testInvalid": map[string]interface{}{
+			"ignored": 1,
+		},
+		"testInvalid2": []interface{}{1, 2},
+	}
+
+	//transform testData into byte slice
+	bitties, _ := json.Marshal(testData)
+
+	if os.Getenv("BadData") == "1" {
 		//create a temp file and write the data to it json encoded
 		f, _ := ioutil.TempFile("", "configTest")
 		filename := f.Name()
 		defer os.Remove(filename)
-		f.Write(bitties)
+		//write in bad data
+		f.Write(bitties[1:])
 		f.Close()
-
 		ReadConfig(filename)
-	}()
+		return
+	}
 
-	//now attempt to read a file that doesn't exist. should panic
-	func() {
-		defer func() {
-			if recover() == nil {
-				T.Errorf("Failed to panic on bad data")
-			}
-		}()
+	if os.Getenv("BadFile") == "1" {
 		ReadConfig("1234567890asdfghjkl")
-	}()
+		return
+	}
+
+	badData := exec.Command(os.Args[0], "-test.run=TestReaderExits")
+	badData.Env = append(os.Environ(), "BadData=1")
+	err := badData.Run()
+	if e, ok := err.(*exec.ExitError); !ok || e.Success() {
+		T.Errorf("Process did not exit")
+	}
+	badFile := exec.Command(os.Args[0], "-test.run=TestReaderExits")
+	badFile.Env = append(os.Environ(), "BadFile=1")
+	err = badFile.Run()
+	if e, ok := err.(*exec.ExitError); !ok || e.Success() {
+		T.Errorf("Process did not exit")
+	}
+
 }
